Add link between absl`s logging and spdlog.
This is just to enable the CHECK_XXX macros to work. Specifically, we
should not try and use absl's logging functionality in general
(although we theoretically could, there is little reason to do so).

This just makes it so that using CHECK macros will give useful
messages in the default log locations (unlike assert-failures, which
only give a stacktrace, these macros can give useful debugging
messages in addition to the stack trace).

Note that due to a quirk in absl's logging library, fatal messages
always show up twice.  I added a silly CHECK_EQ(1, 2) to xid_mgr, and
this is the output now:

  [2025-01-13 19:02:48.663 +00:00] [debug] [xid_partition.cc:150:commit_xid] [thread 588716] Partition: commit xid: db_id=1, xid=10
  [2025-01-13 19:02:48.663 +00:00] [critical] [::] [thread 588716] [/home/dev/springtail/src/xid_mgr/xid_mgr_server.cc:64] Check failed: 1 == 2 (1 vs. 2)
  [2025-01-13 19:02:48.670 +00:00] [critical] [::] [thread 588716] [/home/dev/springtail/src/xid_mgr/xid_mgr_server.cc:64] Check failed: 1 == 2 (1 vs. 2)
  [2025-01-13 19:02:48.910 +00:00] [error] [exception.cc:19:backtrace_handler] [thread 588716] Backtrace from signal 6:
  Stack trace (most recent call first):
  #0  0x0000c89c7b2e215b in backtrace_handler at /home/dev/springtail/src/common/exception.cc:14:47
  ...