Use CHECK_EQ(a, b) rather than assert(a == b) (#218)
This has the advantage of printing out the values of "a" and "b" in
the failure case.

This also has the change that this code will be run all of the time,
even if NDEBUG is set. If the performance overhead of validating these
assertions matters, then one can use DCHECK* macros instead, which
will no-op if NDEBUG is set. In my experience, it is worthwhile to
verify almost all assertions in production environments, unless they are
prohibitively expensive.

In a few cases, code was doing something like
  assert(iter == iter.end())
and in these cases, I did not change to CHECK_EQ, because there is no
way to pretty-print the values.

This came up for me when I was writing a microbenchmark that is
attempting to construct its own MutableBTree. I had a case where I was
using the wrong schema field types. I now get an error like:
  Check failed: this->get_type() == field->get_type() (UINT64 vs. TEXT)
instead of just a notification that they are not equal.

I can take this a step further, and replace every use of "assert" with
"CHECK". There are also CHECK_GT, GE, LT, LE, NE macros, but that is
not done in this change.