Always verify return values of ::read and ::write
This is necessary to make release builds work (or we would need to
disable some warnings).

I worry that someday, these may cause crashes in production, as both
read and write do not guarantee to process all of the bytes, and thus
some of these calls should be wrapped in loops that retry in the case
of partial read/writes. That being said, we can also just wait and see
and fix it when/if it happens.