Add opentelemetry metric for e2e commit latency
This measures the latency from when postgres believes that a
transaction was committed, until we have fully written out the new
btrees to storage (at which point, it is possible for springtail
queries to read the data).

I tested this by running an opentelemetry collector:
  docker pull otel/opentelemetry-collector-contrib:0.118.0
  docker run \
    -p 4317:4317 \
    -p 4318:4318 \
    -p 55679:55679 \
    otel/opentelemetry-collector-contrib:0.118.0 \
    2>&1 | tee collector-output.txt # Optionally tee output for easier search later

And then setting up system.json.test to contain:
    "otel": {
        "enabled": true,
        "host": "172.17.0.3",
        "port": 4317,
        "metrics_host": "172.17.0.3",
        "metrics_port": 4317,
        "metrics_export_interval_millis": 1000,
        "metrics_export_timeout_millis": 1000
    }

And then I did the normal things to test springtail, i.e.
  python3 ./springtail.py -f ../../system.json.test -b ../../debug --start
  (run psql, create table, insert data)
  And then I could search for btree_write_latencies in collector-output.txt and find it