Add example prometheus exporter
This functionality has two possible benefits:
1. It exposes a simple HTTP GET endpoint (/metrics) that can be useful
   when debugging or for our benchmark framework if people want to see
   the values of metrics on a binary right now
2. If we want to use Prometheus (probably + grafana) for monitoring,
   this is one path to enable that.

This change has a substantial downside of requiring an additional port
to be defined for every binary that runs (and that is not currently
plumbed through in the code).

In practice, (2) could also be done by using the OTel collector and
then just point prometheus at the one collector instance (rather than
needing prometheus to somehow discover all of the binaries we are
running). This is probably a better architecture for prometheus than
using this change directly.

For (1), this change could be quite valuable for both manual debugging
during issues, and for possible benchmark visibility/improvements. It
makes an endpoint, http://localhost:8899/metrics, that exports metrics
in a way that prometheus can collect. We could change our benchmarks
to download this URL before/after running tests, and report on
interesting changes if we wanted to.