Modify our grpc port to support opentelemetry, and enable it in the build
I find this change unsatisfying/annoying in many ways, but this is the
best compromise I could come up with.

The problem:

We want to support grpc-opentelemetry: specifically, grpc has lots of
built-in functionality where it will report very useful metrics to
opentelemetry, which can generally make monitoring better.

We would like to support opentelemetry-grpc: gRPC is the most
popular/efficient/common way of reporting both metric values and span
tracing from our binaries to opentelemetry-collector. If we do not use
this, we can use opentelemetry-http instead, which will push these
same values over HTTP instead of gRPC. This is slightly less common,
and probably slightly less efficient, but it does not likely actually
make much difference in practice.

Due to the way that vcpkg works, there is no straightforward way to do
both of the above. It would create a circular dependency to enable
both (as either build requires the previous build to have succeeded
first). I think that it could probably somehow be done by creating a
grpc-and-opentelemetry combined package that somehow built them both
together, but I was not able to figure out an easy way to make that
work. Alternately I think that it could be done in
bigger/scarier/crazier ways if we abandoned vcpkg and directly used
bazel to build these, but that does not seem like a sensible approach
at this point.  Thus, I ended up with this commit - which turns off
opentelemetry-grpc and enabled grpc-opentelemetry.

Note that this change does *not* actually start using the
grpc-opentelemetry monitoring, that will be in a different change -
this change just modifies the vcpkg/dependencies stuff to be built in
a way that this will eventually work.

This change is a functional change, in that the existing
"host"/"port"/"metrics_host"/"metrics_port" options in
system.settings.json now need to refer to HTTP endpoints instead of
gRPC endpoints.  i.e. it should probably use port 4318 (HTTP) rather
than 4317 (gRPC) by default.