Migrate everything from Thrift to gRPC (#281)
* Migrate everything from Thrift to gRPC

The primary motivation for this is to enable opentelemetry tracing to
work across RPC boundaries, which gRPC has automatic support for. A
secondary motivation is that perhaps gRPC is a bit more
mature/documented/etc than Thrift.

Currently xid_manager and write_cache use some amount of PIMPL to hide
complexities from their header files, but sys_tbl_mgr does not. I can
fix this inconsistency if people would like me to.

Instead of doing thrift->json and storing json in redis, we are now
storing serialized protobufs in redis for schema changes.

I felt too weird about how thrift directly put files in
src/include/thrift, so the protobuf generated files are all in the
build directory (instead of the source directory).  This is
potentially an argument for using the PIMPL strategy, as it avoids
needing a PUBLIC target_include_directory to be set, but not a big
deal either way.

I honestly have somewhat mixed feelings about this change. I
personally like gRPC much better than thrift, but I'm not sure that is
a particularly valid reason to migrate. I do think that the
opentelemetry integration will be quite valuable. It will give
automatic counters and latency numbers for all rpc calls I think, and
opentelemetry context will be able to go across RPCs with ease. Doing
the same thing to thrift would not be straightforward, but still might
be a smaller change than this.

For some background on gRPC - a channel is using http/2, and thus one
channel is one TCP connection (possibly SSL) that can multiplex many
RPCs. This *might* be limited to 100 RPCs at once, although when I
look at the source code
./external/vcpkg/buildtrees/grpc/src/v1.68.2-6cbee13a3f.clean/src/core/ext/transport/chttp2/transport/http2_settings.h
I see it set to UINT32MAX or somesuch, so maybe not. At very high RPC
rates though, the single channel can become a bottleneck, and it can
make sense to multiplex over multiple channels.

The gRPC server defaults to synchronous code, but it supports async
modes if we want to run them. In synchronous code, it default to a
maximum number of threads, as defined in grpc_server.hh SetMaxThreads.

The current RPC retry logic is largely guided by max_attempts. In my
experience, to make a system truly resilient, even under extreme
load/latency/etc, it is best to retry infinitely, and to use deadlines
on RPCs where you can tolerate failure. grpc::ClientContext supports
deadlines if we want to add support for them in the future.

This change should make some of the #include compatibility issues with
postgres go away, because we are no longer #defining ERROR to
different values, and not using thrift headers. I also did some
cleanup in include/pg_fdw/pg_fdw_mgr.hh that avoided some postgres
conflicts as well.

Note that when converting from .thrift to .proto, I chose to make
TableInfo.columns be a vector rather than a map. proto3 does support
the map type, but it is an unordered map, unlike thrift which uses
std::map. This also makes TableInfo.columns consistent with
IndexInfo.columns.

* Small fix during merge.

---------

Co-authored-by: Garth Goodson <garth@springtail.io>