New dynamic main (#335)
* Modifications required for dynamic linking on arm64.

* Updates to scripts and cmake to set arm-linux-dynamic triplet.

* Minor change to cmake flags in debug build.

* Catch all SQL execution errors and report them as test failures, refs SPR-502

* Fixes. but fdw doesnt load.

* Create our own port of grpc - this is in preparation for adding otel support to it

This is a direct (unmodified) copy from vcpkg version
6f29f12e82a8293156836ad81cc9bf5af41fe836 (tag 2025.01.13).

* Modify our grpc port to support opentelemetry, and enable it in the build

I find this change unsatisfying/annoying in many ways, but this is the
best compromise I could come up with.

The problem:

We want to support grpc-opentelemetry: specifically, grpc has lots of
built-in functionality where it will report very useful metrics to
opentelemetry, which can generally make monitoring better.

We would like to support opentelemetry-grpc: gRPC is the most
popular/efficient/common way of reporting both metric values and span
tracing from our binaries to opentelemetry-collector. If we do not use
this, we can use opentelemetry-http instead, which will push these
same values over HTTP instead of gRPC. This is slightly less common,
and probably slightly less efficient, but it does not likely actually
make much difference in practice.

Due to the way that vcpkg works, there is no straightforward way to do
both of the above. It would create a circular dependency to enable
both (as either build requires the previous build to have succeeded
first). I think that it could probably somehow be done by creating a
grpc-and-opentelemetry combined package that somehow built them both
together, but I was not able to figure out an easy way to make that
work. Alternately I think that it could be done in
bigger/scarier/crazier ways if we abandoned vcpkg and directly used
bazel to build these, but that does not seem like a sensible approach
at this point.  Thus, I ended up with this commit - which turns off
opentelemetry-grpc and enabled grpc-opentelemetry.

Note that this change does *not* actually start using the
grpc-opentelemetry monitoring, that will be in a different change -
this change just modifies the vcpkg/dependencies stuff to be built in
a way that this will eventually work.

This change is a functional change, in that the existing
"host"/"port"/"metrics_host"/"metrics_port" options in
system.settings.json now need to refer to HTTP endpoints instead of
gRPC endpoints.  i.e. it should probably use port 4318 (HTTP) rather
than 4317 (gRPC) by default.

* enable opentelemetry-prometheus support

This is to have all of the dependency changes made at once, so that we
don't need to keep making slow developer builds and upgrading docker
images.

This gives the option of exposing prometheus HTTP endpoints on our
binaries, which could be good for debugging and scripting. See
https://github.com/Springtail-inc/springtail/pull/288 for more
context.

* Add grpc-otel support

This automatically records opentelemetry metrics (i.e. counts, error
counts?, latencies, payload size distribution, approx total bandwidth)
for grpc server/client RPCs.

* grpc_client: add opentelemetry tracing support

This adds trace spans for client ops, and also sends grpc span info
along to the server.

* grpc_server: add otel tracing spans for server side

I spent a while investigating if I could do this with grpc Interceptor
support instead, but I couldn't get it to a place that I was happy
with. This way seemed more explicit and potentially better in any
case.

* Improve otel logging and add service name to all traces

* fix release mode build error

* Fix logging for OTel with HTTP

* Working on dynamic linking

* Remove unused ports and fix a errno issue in pg_repl_connection

* Fix issues with merge and move around gRPC code

* Fix a couple of import issues

* Remove the http link, the protocol value can be passed from config

* Fix tracing.cc headers and minor review comments

* Adding environment file for handling shared library loading in postgres

* Fixup issues post merge

* Fix from Sonarqube

* Some fixes for sonarqube.

* Add back spans in XidMgr, Use template for GrpcServer

* Fix hardcoded path

* Fix the CMakeLists.txt to use the Project Source Dir for the vcpkg-ports

* Fix to vcpkg.json

---------

Co-authored-by: Craig Soules <craig@springtail.io>
Co-authored-by: Cliff Frey <cliff@springtail.io>
Co-authored-by: Arun <arun@springtail.io>